appium/iOS/appium_cerra_ios_22nd_March/appium_cerra_ios/requirements.txt

#A
Appium-Python-Client==4.0.0
allure-pytest==2.13.3


#N
numpy==1.26.4


#O
openpyxl==3.1.2
opencv-python==4.9.0.80


#P
pytest==8.1.1
pandas==2.2.1


#r
requests==2.31.0
